requirements.txt
requests>=2.25.1
pyyaml>=5.4.1
selenium>=3.141.0